E2E-testing with puppeteer (#4)
* Setup E2E-testing with puppeteer

* Fixed build-failure

From https://www.typescriptlang.org/tsconfig#types
```
By default all visible ”@types” packages are included in your compilation.
If types is specified, only packages listed will be included.
Other packages under node_modules/@types/* will not be included.
```

So if any of the other packages has a bug, it will blow up my build. Hence exclude all. If one type is needed, include it later.

* CI for e2e-testing

* Create junit.xml

* Increased testTimeout

* server-package as test-server for E2E-testing

Moved jest-puppeteer.config.js to project-root, as things are a lot easier, than being in test/e2e-folder. So far it was OK, but with starting the server it got complicated. So move it.

* created some appTest

* Updated CI

* Use caching in e2e.yml