Show hint Vue base element
If Vue can't start, then the element won't be replaced, thus showing this hint.
Downside is that this hint is shown always before Vue kicks in. On fast networks this isn't a problem, but on slow ones it's a bit misleading, as Vue can work so the hint gives a wrong clue.