Hint for old browsers (#20)
* Show hint Vue base element

If Vue can't start, then the element won't be replaced, thus showing this hint.
Downside is that this hint is shown always before Vue kicks in. On fast networks this isn't a problem, but on slow ones it's a bit misleading, as Vue can work so the hint gives a wrong clue.

* Use global appLoaded and listen to window error-events

If the MainView is mounted without error, the flag is set thus indicating successful load.
In case of any errors the event-handler is invoked, which displays notSupportedBrowserInfo.

Downside:
* any error will display this div, even if unrelated to old browser.
* if there is anywhere in the app some other feature that the browser doesn't support this won't be caught here

* Test for SyntaxError

* Revert "Test for SyntaxError"

This reverts commit f4c1ee6dbe02724311e12364b29d55af0c6f3d8d.
It won't work, because some browser use a localized text for "SyntaxError"

* Use browserlists to show the hint

If Vue can start, it will remove the hint. The hint states "unsupported", so it may run or not.
I think this is an adequate solution.